Develop

+++ src/main/java/com/group1/quiz/service/UserService.java
@@ -1,9 +1,11 @@
 package com.group1.quiz.service;
 
 import com.group1.quiz.dataTransferObject.UserDto;
-import com.group1.quiz.model.UserRole;
+import com.group1.quiz.enums.UserRoleEnum;
 import com.group1.quiz.model.UserModel;
 import com.group1.quiz.repository.UserRepository;
+import java.time.Instant;
+import java.util.Date;
 import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -22,33 +24,66 @@ public class UserService implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(String username) {
         Optional<UserModel> userModel = userRepository.findUserByUsername(username);
-        String defaultUsername = "Parzival";
-        String defaultPassword = new BCryptPasswordEncoder().encode("123");
+        String defaultUsername = "string";
+        String defaultPassword = new BCryptPasswordEncoder().encode("string");
 
         if(userModel.isPresent()) {
             return org.springframework.security.core.userdetails.User.builder()
                     .username(userModel.get().getUsername())
                     .password(userModel.get().getPassword())
-                    .roles(String.valueOf(userModel.get().getRole()))
+                    .roles(userModel.get().getRole().getValue())
                     .build();
         }
         else if(Objects.equals(username, defaultUsername)) {
             return org.springframework.security.core.userdetails.User.builder()
                     .username(defaultUsername)
                     .password(defaultPassword)
-                    .roles(String.valueOf(UserRole.ADMIN))
+                    .roles(UserRoleEnum.ADMIN.getValue())
                     .build();
         }
         return null;
     }
-    public void createUser(UserDto userDto){
+    public void createUser(UserDto userDto) throws Exception {
+        if(userRepository.findUserByUsername(userDto.getUsername()).isPresent()) {
+            throw new Exception("Username already exists");
+        }
         UserModel userModel = new UserModel(userDto);
         String encodedPassword = new BCryptPasswordEncoder().encode(userModel.getPassword());
         userModel.setPassword(encodedPassword);
-        userRepository.save(userModel);
+        userRepository.insert(userModel);
     }
 
-    public void deleteUser(String id) {
-        userRepository.deleteById(id);
+
+    public void updateUser(String id, UserDto userDto) throws Exception {
+        Optional<UserModel> userModel = userRepository.findById(id);
+        if(userModel.isPresent()) {
+            UserModel user = userMapping(id, userDto);
+            userRepository.save(user);
+        }
+        else {
+            throw new Exception("User not found");
+        }
     }
+
+    private UserModel userMapping(String id, UserDto userDto) {
+        return UserModel.builder()
+                .id(id)
+                .username(userDto.getUsername())
+                .password(new BCryptPasswordEncoder().encode(userDto.getPassword()))
+                .role(userDto.getRole())
+                .createdAt(Date.from(Instant.now()))
+                .updatedAt(Date.from(Instant.now()))
+                .build() ;
+    }
+
+    public void deleteUser(String id) throws Exception {
+        Optional<UserModel> userModel = userRepository.findById(id);
+        if(userModel.isPresent()) {
+            userRepository.deleteById(id);
+        }
+        else {
+            throw new Exception("User not found");
+        }
+    }
+
 }
